refactored tx.get
